requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
jinja2==3.1.2
playwright==1.40.0
redis==5.0.1 